requirements.txt
stable-ts==2.13.3
pydub==0.25.1
python-docx==1.1.0
pyyaml==6.0.2
librosa==0.10.2.post1
soundfile==0.12.1
streamlit==1.39.0
pyarrow==17.0.0
